Handle types module imports in a more sane way
