+++ chatlas/_utils.py
@@ -64,11 +64,22 @@ def is_testing():
     return os.environ.get("PYTEST_CURRENT_TEST", None) is not None
 
 
+class MISSING_TYPE:
+    """
+    A singleton representing a missing value.
+    """
+
+    pass
+
+
+MISSING = MISSING_TYPE()
+
+
 class DefaultModelWarning(Warning):
     pass
 
 
-def inform_model_default(model: str, stacklevel: int = 3  ) -> str:
+def inform_model_default(model: str, stacklevel: int = 3) -> str:
     if not is_testing():
         msg = f"Defaulting to `model = '{model}'`."
         warnings.warn(msg, DefaultModelWarning, stacklevel=stacklevel)
